Newspaper API/requirements.txt
nltk
python-dotenv
newspaper3k
tkinter